Rearrange CommentsApi into room.ts
